AMD64 - Fix bogus signal interruption
* The TF_XFLAGS field in the trap frame was not getting cleared on trap and
  syscall entry.  This field can cause a mailbox signal to be flagged in
  the signal frame which in turn can cause the next blockable system call
  to return EINTR.

  This fixes the last issue with mount_mfs.

* Cast the signal function to intptr_t rather then int.

+++ sys/kern/kern_sig.c
@@ -1677,8 +1677,8 @@ issignal(struct lwp *lp, int maytrace)
 		 * Return the signal's number, or fall through
 		 * to clear it from the pending mask.
 		 */
-		switch ((int)(intptr_t)p->p_sigacts->ps_sigact[_SIG_IDX(sig)]) {
-		case (int)SIG_DFL:
+		switch ((intptr_t)p->p_sigacts->ps_sigact[_SIG_IDX(sig)]) {
+		case (intptr_t)SIG_DFL:
 			/*
 			 * Don't take default actions on system processes.
 			 */
@@ -1731,7 +1731,7 @@ issignal(struct lwp *lp, int maytrace)
 
 			/*NOTREACHED*/
 
-		case (int)SIG_IGN:
+		case (intptr_t)SIG_IGN:
 			/*
 			 * Masking above should prevent us ever trying
 			 * to take action on an ignored signal other
